Move state delta handler to separate module
This will make room for a blockchain submission module.
